backend/requirements.txt
django==3.2.12
djangorestframework
mysqlclient
django-mysql
drf-writable-nested
